DOCKER-557: fast `docker rm`
Do not wait until a VM is actually destroyed/deleted to return from a
DELETE /vms/vm_uuid request. Instead, stop the VM and set its property
`transitive_state` to `'destroying'`.

At that point, do not allow any change to this VM except for marking it
as destroyed when it's actually been destroyed.

This change comes with a script named `add-transitive-state-index.js` to
add an index on `transitive_state` to moray.